requirements.txt
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.8.3
cffi==2.0.0
colorama==0.4.6
distro==1.9.0
exceptiongroup==1.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.11.0
numpy==2.2.6
openai==1.108.1
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
python-dotenv==1.1.1
sniffio==1.3.1
sounddevice==0.5.2
soundfile==0.13.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
websockets==15.0.1
